Menage dans les fichiers + Avancement sur le parseur ( non testé (mettre en place des TU avec GoogleTest ou cppunit) + crea des classes profil,Admin, Droits(Enum), UUID( https://stackoverflow.com/questions/24365331/how-can-i-generate-uuid-in-c-without-using-boost-library )

+++ FrameworkQT/src/main/c++/fr/univtours/polytech/FrameworkQT/model/Utilsateur.cpp
@@ -1,8 +1,9 @@
 #include "Utilisateur.h"
 
-unsigned long int Utilisateur::getId() {
+int Utilisateur::getId() {
     return id;
 }
+
 void Utilisateur::setId(unsigned long int newId) {
     id = newId;
 }
